+++ CHANGELOG.md
@@ -43,7 +43,7 @@
 
 ### Features
 
-* New AI level 6, docs updated ([23b79f2](https://github.com/josefjadrny/js-chess-engine/commit/23b79f28fc35e9a667c50d3bf237fcf14a4b8c9a))
+* New highest AI difficulty level added, docs updated ([23b79f2](https://github.com/josefjadrny/js-chess-engine/commit/23b79f28fc35e9a667c50d3bf237fcf14a4b8c9a))